requirements.txt
edge-tts
python-dotenv
google-cloud-texttospeech
